update
Make the underline works.
Adding it forEach sections and menu-links .
Where i store if statements where for sections i checks it if window.scroll position is offset from Top each section and if its i store section id to variable. Next i check if menulinks includes href atribute of current section (current section id ) and also checking if currentsection is not empty, because for not showing underline at top of the page (homepage). If its true adding a classlist. active