update
recreating the order form afte submiting.
The form information is store to the variable newOrderData and then with for loop they are show to the console log.
Replace the form with the div that thanks for order.
SetTimeout function for closing the dialog, reseting the form and make a form to original form content after 5 seconds